Remove `name` check from `LookupConstructors`
This parameter required the unscoped type name, but CppInterOp provides
no way of getting that. Every usage of this appears to have a hard-coded
type name instead. After speaking with Vassil on Discord, he suggested
this additional check may not actually be needed.

For now, the param has been kept, to not break existing code. If
everything works correctly, we can follow up to remove the param
entirely.

If this param cannot be removed, please provide a way for the unscoped
name to be accessed via CppInterOp.