Remove dynamic allocation path from ctor wrapper
This is a very surprising bit of codegen, which uses `new` to
dynamically allocate a new object if a ctor is called without an
invoking object. This is not a safe assumption, adds a new branch to
every ctor call, and also doesn't allow for custom allocators.
Altogether, the code was doing too much.

If someone needs a new object, let them allocate it first and then run
the ctor on it.

ALSO, this commit removes one layer of indirection, where the invoking
object was always coming in as a pointer to pointer. This isn't needed
and it's not how args are handled. Placement new just needs a single
pointer, so I've removed the expectation of additional indirection.
Keeping this makes LLVM IR gen to work with this function much more
complex.