Patch call wrapper to not check ret
Previously, for a non-void function, we would generate code which
checks if `ret` is set. If it is, we'd do a placement new on it,
otherwise, we'd treat the function as void. This changes the behavior to
always require `ret` to be set.

This check was an arbitrary convenience, at the cost of twice as much
C++ generated per call and an additional runtime branch.