Add AOT call support
This adds a new `AotCall` type, similar to `JitCall`, except it has a
partial TU which can be used for AOT compilation. There is a
new `MakeAotCallable` function which works on function, field, and var
decls.

The existing wrapper generators have been turned into templates so the
code can be reused across both AotCalls and JitCalls.